refactor(palletizing): expose Harmonic implementation as HAL
