+++ exercises/palletizing/python_template/HAL.py
@@ -1,59 +1,167 @@
-print("HAL initializing", flush=True)
-##############################################################################
-# JdeROBOT ROBOTICS ACADEMY (http://jderobot.github.io/RoboticsAcademy/)
-#  API PICK and PLACE exercise, including:
-#   Robot Info: get_TCP_pose, get_Joint_states
-#   Kinematics: MoveAbsJ, MoveJoint, MoveLinear, MoveSingleJ
-#               MoveRelLinear, MoveRelReor
-#   Gripper: GripperSet, attach, dettach
-#
-#   VERSION: 1.0
-# 	DATE: 	 April 21, 2025
-#   AUTHOR:  Diego Martin (diego.martin.martin@gmail.com)
-#
-# ======= Acknowledgments =======
-#  IFRA-Cranfield nice "ROS2 Sim-to-Real Robot Control" package
-#  URL: https://github.com/IFRA-Cranfield/ros2_SimRealRobotControl
-##############################################################################
+print("HAL Harmonic initializing", flush=True)
 
-import sys, os, time, math
+import sys, os, time, math, json, copy
 import rclpy
-
-# import tf_transformations
 import numpy as np
+from sensor_msgs.msg import JointState
 
 from rclpy.node import Node
-from sensor_msgs.msg import JointState
 from ros2srrc_data.msg import Robpose
-from linkattacher_msgs.srv import AttachLink, DetachLink
 from ament_index_python.packages import get_package_share_directory
 
-# Build PATH and import Python classes from IFRA package:
+from std_msgs.msg import Bool, String
+
+# Paths
 PATH = os.path.join(get_package_share_directory("ros2srrc_execution"), "python")
 
-PATH_ROB = PATH + "/robot"
-sys.path.append(PATH_ROB)
+sys.path.append(PATH + "/robot")
 from robot import RBT
 
-PATH_EE = PATH + "/endeffector"
-sys.path.append(PATH_EE)
-from robotiq_ur import RobotiqGRIPPER
+# ROS msgs
+from ros2srrc_data.msg import Action, Joint, Joints, Xyz, Ypr
 
-# Import ROS2 Custom Messages from IFRA package:
-from ros2srrc_data.msg import Action
-from ros2srrc_data.msg import Joint
-from ros2srrc_data.msg import Joints
-from ros2srrc_data.msg import Xyz
-from ros2srrc_data.msg import Ypr
-from ros2srrc_data.msg import Robpose
-
-# Inicialization
+# Init
 rclpy.init(args=None)
 UR5 = RBT()
 
+HAL = Node("hal_node")
+HAL.grasped_object = None
 
-#################################### ROBOT KINEMATICS ###################################################
-# MoveAbsJ. Absolute Joints in degrees, speed max 1.0, wait time after movement in seconds
+HAL.auto_attach_pub = HAL.create_publisher(
+    Bool,
+    "/gripper_auto_attach",
+    10,
+)
+
+HAL.graspable_pub = HAL.create_publisher(
+    String,
+    "/graspable_objects",
+    10,
+)
+
+# The attacher uses substring matching, so "box" covers feeder-generated names.
+GRASPABLE_OBJECTS = "box"
+
+graspable_msg = String()
+graspable_msg.data = GRASPABLE_OBJECTS
+HAL.graspable_pub.publish(graspable_msg)
+
+print("[HAL] Published graspable objects")
+
+
+def publish_graspable_objects():
+
+    graspable_msg = String()
+    graspable_msg.data = GRASPABLE_OBJECTS
+    HAL.graspable_pub.publish(graspable_msg)
+
+
+# Periodic republish so late-joining subscribers receive the list
+HAL.create_timer(1.0, publish_graspable_objects)
+
+# /box_ready announces a stopped box. /box_done acknowledges that it has been
+# lifted clear so the feeder can advance. ROS topics remain hidden by the HAL.
+
+HAL._ready_box = None
+HAL._processed_boxes = set()
+HAL._box_infos = {}
+HAL._pickup_poses = {}
+HAL._pallet_info = None
+
+
+def _on_box_ready(msg):
+    # Ignore re-announcements of a box we've already handed to the solution.
+    if msg.data not in HAL._processed_boxes:
+        HAL._ready_box = msg.data
+
+
+HAL.create_subscription(String, "/box_ready", _on_box_ready, 10)
+HAL.box_done_pub = HAL.create_publisher(String, "/box_done", 10)
+
+
+def _on_box_info(msg):
+    try:
+        info = json.loads(msg.data)
+        name = info["name"]
+        pickup_pose = info["pickup_pose"]
+        if pickup_pose.get("frame") != "base_link":
+            raise ValueError("pickup_pose must use the base_link frame")
+        if len(pickup_pose["center"]) != 3:
+            raise ValueError("pickup_pose center must contain x, y, z")
+
+        HAL._box_infos[name] = {
+            "name": name,
+            "sku": info["sku"],
+            "size": info["size"],
+            "mass": info["mass"],
+        }
+        HAL._pickup_poses[name] = pickup_pose
+    except (json.JSONDecodeError, KeyError, TypeError, ValueError) as exc:
+        print(f"[HAL] Ignoring invalid /box_info message: {exc}")
+
+
+def _on_pallet_info(msg):
+    try:
+        HAL._pallet_info = json.loads(msg.data)
+    except json.JSONDecodeError as exc:
+        print(f"[HAL] Ignoring invalid /pallet_info JSON: {exc}")
+
+
+HAL.create_subscription(String, "/box_info", _on_box_info, 10)
+HAL.create_subscription(String, "/pallet_info", _on_pallet_info, 10)
+
+
+def WaitForBox():
+    """Wait for a box at the pickup point and return its feeder name."""
+    while rclpy.ok() and HAL._ready_box is None:
+        rclpy.spin_once(HAL, timeout_sec=0.1)
+    name = HAL._ready_box
+    HAL._ready_box = None
+    HAL._processed_boxes.add(name)
+    return name
+
+
+def GetBoxInfo(name, timeout=5.0):
+    """Return semantic task metadata for a box announced by WaitForBox()."""
+    deadline = time.time() + timeout
+    while rclpy.ok() and name not in HAL._box_infos and time.time() < deadline:
+        rclpy.spin_once(HAL, timeout_sec=0.1)
+    if name not in HAL._box_infos:
+        raise RuntimeError(f"No semantic /box_info received for {name}")
+    return copy.deepcopy(HAL._box_infos[name])
+
+
+def GetPickupPose(name, timeout=5.0):
+    """Return the observed box top-center pose in the base_link frame."""
+    deadline = time.time() + timeout
+    while rclpy.ok() and name not in HAL._pickup_poses and time.time() < deadline:
+        rclpy.spin_once(HAL, timeout_sec=0.1)
+    if name not in HAL._pickup_poses:
+        raise RuntimeError(f"No pickup pose received for {name}")
+    return copy.deepcopy(HAL._pickup_poses[name])
+
+
+def GetPalletInfo(timeout=5.0):
+    """Return pallet metadata in the robot base_link frame."""
+    deadline = time.time() + timeout
+    while rclpy.ok() and HAL._pallet_info is None and time.time() < deadline:
+        rclpy.spin_once(HAL, timeout_sec=0.1)
+    if HAL._pallet_info is None:
+        raise RuntimeError("No /pallet_info received")
+    return copy.deepcopy(HAL._pallet_info)
+
+
+def BoxDone(name):
+    """Acknowledge that the box is clear so the feeder can advance."""
+    HAL.box_done_pub.publish(String(data=name))
+    print(f"[HAL] BoxDone({name}) -> released next box")
+
+
+# ==============================================================
+# MoveAbsJ (IDÉNTICO a classic)
+# ==============================================================
+
+
 def MoveAbsJ(absolute_joints, speed, wait_time):
 
     ACTION = Action()
@@ -71,8 +179,7 @@ def MoveAbsJ(absolute_joints, speed, wait_time):
 
     EXECUTION = UR5.Move_EXECUTE(ACTION)
 
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
+    if EXECUTION["Success"]:
         print(f"Robot moved to Joint Angular Goal: {absolute_joints}")
         print(
             f"Movement Execution Time: {EXECUTION['ExecTime']} s at Robot Speed: {speed*100} %"
@@ -80,14 +187,15 @@ def MoveAbsJ(absolute_joints, speed, wait_time):
     else:
         print("Robot movement FAILED, check REASON in MoveIt output")
 
-    # Wait till next movement
     time.sleep(wait_time)
-    print(f"Waiting {wait_time} s")
-    print("")
+    print(f"Waiting {wait_time} s\n")
 
 
-# MoveLinear. Linear movement to absolute pose XYZ with YPR absolute orientation in degrees
-# Speed max 1.0, wait time after movement in seconds
+# ==============================================================
+# MoveLinear
+# ==============================================================
+
+
 def MoveLinear(abs_xyz, abs_ypr, speed, wait_time):
 
     roll = math.radians(abs_ypr[0])  # Converts XYR to rad
@@ -134,15 +242,17 @@ def MoveLinear(abs_xyz, abs_ypr, speed, wait_time):
     print("")
 
 
-# MoveJoint. Point-to-point movement to absolute pose XYZ with YPR absolute orientation in degrees
-# Speed max 1.0, wait time after movement in seconds
+# ==============================================================
+# MoveJoint
+# ==============================================================
+
+
 def MoveJoint(abs_xyz, abs_ypr, speed, wait_time):
 
-    roll = math.radians(abs_ypr[0])  # Converts XYR to rad
+    roll = math.radians(abs_ypr[0])
     pitch = math.radians(abs_ypr[1])
     yaw = math.radians(abs_ypr[2])
 
-    # Quaternion from YPT in rad
     qx = np.sin(roll / 2) * np.cos(pitch / 2) * np.cos(yaw / 2) - np.cos(
         roll / 2
     ) * np.sin(pitch / 2) * np.sin(yaw / 2)
@@ -167,8 +277,7 @@ def MoveJoint(abs_xyz, abs_ypr, speed, wait_time):
 
     EXECUTION = UR5.RobMove_EXECUTE("PTP", float(speed), InputPose)
 
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
+    if EXECUTION["Success"]:
         print(
             f"Robot moved Point-to-Point to Abs XYZ: {abs_xyz} and Abs YPR: {abs_ypr}"
         )
@@ -178,15 +287,42 @@ def MoveJoint(abs_xyz, abs_ypr, speed, wait_time):
     else:
         print("Robot movement FAILED, check REASON in MoveIt output")
 
-    # Wait till next movement
     time.sleep(wait_time)
-    print(f"Waiting {wait_time} s")
-    print("")
+    print(f"Waiting {wait_time} s\n")
 
 
-# MoveRelLinear. Linear movement, relative cartesian coordinates
-# Speed max 1.0, wait time after movement in seconds
+# ==============================================================
+# MoveRelLinear
+# ==============================================================
+
+
+def _wait_motion_complete(timeout=15.0, vel_threshold=0.01, stable_count=8):
+    """Block until all joint velocities stay below threshold for stable_count cycles."""
+    latest = [None]
+
+    sub = HAL.create_subscription(
+        JointState, "/joint_states", lambda msg: latest.__setitem__(0, msg), 10
+    )
+    stable = 0
+    deadline = time.time() + timeout
+    try:
+        while time.time() < deadline:
+            rclpy.spin_once(HAL, timeout_sec=0.05)
+            msg = latest[0]
+            if msg is not None and len(msg.velocity) > 0:
+                if all(abs(v) < vel_threshold for v in msg.velocity):
+                    stable += 1
+                    if stable >= stable_count:
+                        return True
+                else:
+                    stable = 0
+    finally:
+        HAL.destroy_subscription(sub)
+    return False
+
+
 def MoveRelLinear(relative_xyz, speed, wait_time):
+
     ACTION = Action()
     ACTION.action = "MoveL"
     ACTION.speed = float(speed)
@@ -199,47 +335,26 @@ def MoveRelLinear(relative_xyz, speed, wait_time):
 
     EXECUTION = UR5.Move_EXECUTE(ACTION)
 
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
+    if EXECUTION["Success"]:
         print(f"Robot moved LINEARLY by a relative increment of : {relative_xyz}")
         print(
             f"Movement Execution Time: {EXECUTION['ExecTime']} s at Robot Speed: {speed*100} %"
         )
+        _wait_motion_complete()  # block until joint velocities settle
     else:
         print("Robot movement FAILED, check REASON in MoveIt output")
 
-
-# MoveSingleJ. Relative angle in degrees, speed max 1.0, wait time after movement in seconds
-def MoveSingleJ(joint_number, relative_angle, speed, wait_time):
-    ACTION = Action()
-    ACTION.action = "MoveR"
-    ACTION.speed = float(speed)
-
-    INPUT = Joint()
-    INPUT.joint = str(joint_number)
-    INPUT.value = float(relative_angle)
-    ACTION.mover = INPUT
-
-    EXECUTION = UR5.Move_EXECUTE(ACTION)
-
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
-        print(f"Robot moved {joint_number} in {relative_angle} degrees")
-        print(
-            f"Movement Execution Time: {EXECUTION['ExecTime']} s at Robot Speed: {speed*100} %"
-        )
-    else:
-        print("Robot movement FAILED, check REASON in MoveIt output")
-
-    # Wait till next movement
     time.sleep(wait_time)
-    print(f"Waiting {wait_time} s")
-    print("")
+    print(f"Waiting {wait_time} s\n")
 
 
-# Relative Reorient given relative Euler Angles
-# Speed max 1.0, wait time after movement in seconds
+# ==============================================================
+# MoveRelReor
+# ==============================================================
+
+
 def MoveRelReor(relative_ypr, speed, wait_time):
+
     ACTION = Action()
     ACTION.action = "MoveROT"
     ACTION.speed = float(speed)
@@ -252,8 +367,7 @@ def MoveRelReor(relative_ypr, speed, wait_time):
 
     EXECUTION = UR5.Move_EXECUTE(ACTION)
 
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
+    if EXECUTION["Success"]:
         print(f"TCP reoriented by a relative increment of : {relative_ypr}")
         print(
             f"Movement Execution Time: {EXECUTION['ExecTime']} s at Robot Speed: {speed*100} %"
@@ -261,128 +375,67 @@ def MoveRelReor(relative_ypr, speed, wait_time):
     else:
         print("Robot movement FAILED, check REASON in MoveIt output")
 
-    # Wait till next movement
     time.sleep(wait_time)
-    print(f"Waiting {wait_time} s")
-    print("")
+    print(f"Waiting {wait_time} s\n")
 
 
-###################################### GRIPPER ###################################################
-class LinkAttacherClient(Node):
-    def __init__(self):
-        super().__init__("link_attacher_client")
-        self.attach_client = self.create_client(AttachLink, "/ATTACHLINK")
-        self.detach_client = self.create_client(DetachLink, "/DETACHLINK")
-
-        while not self.attach_client.wait_for_service(timeout_sec=1.0):
-            self.get_logger().info("Attach service not available, waiting again...")
-        while not self.detach_client.wait_for_service(timeout_sec=1.0):
-            self.get_logger().info("Detach service not available, waiting again...")
-
-    def send_attach_request(self, model1_name, link1_name, model2_name, link2_name):
-        request = AttachLink.Request()
-        request.model1_name = model1_name
-        request.link1_name = link1_name
-        request.model2_name = model2_name
-        request.link2_name = link2_name
-
-        future = self.attach_client.call_async(request)
-        rclpy.spin_until_future_complete(self, future)
-        return future.result()
-
-    def send_detach_request(self, model1_name, link1_name, model2_name, link2_name):
-        request = DetachLink.Request()
-        request.model1_name = model1_name
-        request.link1_name = link1_name
-        request.model2_name = model2_name
-        request.link2_name = link2_name
-
-        future = self.detach_client.call_async(request)
-        rclpy.spin_until_future_complete(self, future)
-        return future.result()
+# ==============================================================
+# MoveSingleJ
+# ==============================================================
 
 
-# Attach object to gripper. Must be called explicitely
-def attach(item):
-    link_attacher_client = LinkAttacherClient()
+def MoveSingleJ(joint_number, relative_angle, speed, wait_time):
 
-    # Attach service call
-    attach_response = link_attacher_client.send_attach_request(
-        "ur5", "EE_robotiq_2f85", item, item
-    )
-    link_attacher_client.get_logger().info(
-        "Attach Response: %s" % attach_response.success
-    )
-
-
-# Dettach all objects. It is always called when gripper is set to full open (0%)
-def dettach():
-    link_attacher_client = LinkAttacherClient()
-
-    # Detach operation for all possible objects when gripper is set to 0%
-    link_attacher_client.send_detach_request(
-        "ur5", "EE_robotiq_2f85", "red_box", "red_box"
-    )
-    link_attacher_client.send_detach_request(
-        "ur5", "EE_robotiq_2f85", "yellow_box", "yellow_box"
-    )
-    link_attacher_client.send_detach_request(
-        "ur5", "EE_robotiq_2f85", "blue_sphere", "blue_sphere"
-    )
-    link_attacher_client.send_detach_request(
-        "ur5", "EE_robotiq_2f85", "green_cylinder", "green_cylinder"
-    )
-
-
-# Gripper closing and opeining to a given percentage (100% full open, 0% full closed)
-# Speed max 1.0, wait time after movement in seconds
-def GripperSet(relative_closure, wait_time):
     ACTION = Action()
-    ACTION.action = "MoveG"
-    ACTION.speed = float(1)  # Gripper speed not working for Robotiq 85, set to 100%
+    ACTION.action = "MoveR"
+    ACTION.speed = float(speed)
 
-    ACTION.moveg = float(relative_closure)
+    INPUT = Joint()
+    INPUT.joint = str(joint_number)
+    INPUT.value = float(relative_angle)
+    ACTION.mover = INPUT
 
     EXECUTION = UR5.Move_EXECUTE(ACTION)
 
-    # Print movement results if movement succeeded
-    if EXECUTION["Success"] == True:
-        print(f"Gripper set to a percentage of: {relative_closure} %")
-        print(f"Movement Execution Time: {EXECUTION['ExecTime']} s")
-        if relative_closure == 0:
-            dettach()  # Automatic object dettach from gripper when full open (0%)
-
+    if EXECUTION["Success"]:
+        print(f"Robot moved {joint_number} in {relative_angle} degrees")
+        print(
+            f"Movement Execution Time: {EXECUTION['ExecTime']} s at Robot Speed: {speed*100} %"
+        )
     else:
-        print("Gripper closing FAILED, check REASON in MoveIt output")
+        print("Robot movement FAILED, check REASON in MoveIt output")
 
-    # Wait till next movement
     time.sleep(wait_time)
-    print(f"Waiting {wait_time} s")
-    print("")
+    print(f"Waiting {wait_time} s\n")
 
 
-from std_msgs.msg import Bool, String
+# ==============================================================
+# SUCTION GRIPPER CONTROL
+# ==============================================================
 
 
-class SuctionNode(Node):
-    def __init__(self):
-        super().__init__("suction_client")
-        self.pub_attach = self.create_publisher(Bool, "/gripper_auto_attach", 10)
-        self.pub_obj = self.create_publisher(String, "/graspable_objects", 10)
+def SuctionSet(on, wait_time):
+    """
+    Turn the suction (vacuum) gripper on or off.
 
-    def set_suction(self, state):
-        msg_obj = String()
-        msg_obj.data = "box"
-        self.pub_obj.publish(msg_obj)
-        msg = Bool()
-        msg.data = state
-        self.pub_attach.publish(msg)
+    on = True   -> grip: any graspable object touching the cup is attached
+    on = False  -> release: the held object is detached
 
+    The suction cup has no actuated joint and no controller. Gripping is handled
+    by the gz_link_attacher plugin, which rigidly attaches an object to the cup
+    link on contact while auto-attach is enabled and releases it when disabled.
+    """
 
-suction_node = SuctionNode()
+    print("\n==================================================")
+    print(f"[HAL] SuctionSet({on}) called")
 
+    auto_msg = Bool()
+    auto_msg.data = bool(on)
+    HAL.auto_attach_pub.publish(auto_msg)
 
-def SuctionSet(state, wait_time):
-    suction_node.set_suction(state)
+    print(f"[HAL] Suction {'ENABLED' if on else 'DISABLED'}")
+
     time.sleep(wait_time)
-    print(f"Suction set to {state}")
+
+    print(f"[HAL] Waiting {wait_time} s")
+    print("==================================================\n")
